% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma.R
\name{bma}
\alias{bma}
\title{Bayesian model averaging for SUGS VarSel}
\usage{
bma(X, MLnorm, Result)
}
\arguments{
\item{X}{Data matrix with observations as rows}

\item{MLnorm}{A vector of normalised marginal likelihood, most likely computed from \code{occamsWindow}}

\item{Result}{The output from the function \code{sugsvarsel}}
}
\value{
Returns a numerical bayesian model averaged coclustering matrix
and numerical vector of bayesian model averaged selected variables
}
\description{
Bayesian model averaging for SUGS VarSel
}
