% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compMlvarsel.R
\name{compMlvarsel}
\alias{compMlvarsel}
\title{Computes the log marginal likelihood of clustering and variable selection}
\usage{
compMlvarsel(K, D, n, nu_0, S_0, lambda_0, nu, lambda, S, lognullMarg,
  intfeature)
}
\arguments{
\item{K}{The number of currently occupied clusters}

\item{D}{The number of variable in the data matrix.}

\item{n}{The vector indicating the number of observations in each cluster}

\item{nu_0}{The degrees of freedom hyperparameter, the default value is \code{D}, where \code{D} is the number of variables.}

\item{S_0}{The scale hyperparamter, the deault value is a fifth of the column variance of the data matrix.}

\item{lambda_0}{The variance of the Guassian mean prior, the dafault value is \code{0.01}.}

\item{nu}{The current posterior degrees of freedom}

\item{lambda}{The current posterior mean variance}

\item{S}{The current posterior scale vector}

\item{lognullMarg}{The log marginal probability of a variable belonging to the null model}

\item{intfeature}{A binary vector of feature which are parition as irrelevant (0)
or relevant (1).}
}
\value{
The log marginal likelhood
}
\description{
Computes the log marginal likelihood of clustering and variable selection
}
