% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compPmlvarsel.R
\name{compPmlvarsel}
\alias{compPmlvarsel}
\title{Computes the log PML for SUGS VarSel in the case of Gaussian mixtures}
\usage{
compPmlvarsel(X, K, N, D, n, phi, betaHat, m, nu, lambda, S, mu_0, nu_0,
  lambda_0, S_0, intfeature, w)
}
\arguments{
\item{X}{The data matrix with observations as rows}

\item{K}{The number of currently occupied clusters}

\item{N}{The number of observations}

\item{D}{The number of variables}

\item{n}{The vector indicating the number of observations in each cluster}

\item{phi}{Current value of posterior dirichlet weights.}

\item{betaHat}{A grid of hyperparameters for the dirichlet concentration parameter, the default is \code{c(0.01, 0.1, 1, 5, 10, 15, 30, 50, 100)}.}

\item{m}{The current posterior mean}

\item{nu}{The current posterior degrees of freedom}

\item{lambda}{The current posterior mean variance}

\item{S}{The current posterior scale vector}

\item{mu_0}{The mean hyperparameter, default is the column means of the data matrix.}

\item{nu_0}{The degrees of freedom hyperparameter, the default value is \code{D}, where \code{D} is the number of variables.}

\item{lambda_0}{The variance of the Guassian mean prior, the dafault value is \code{0.01}.}

\item{S_0}{The scale hyperparamter, the deault value is a fifth of the column variance of the data matrix.}

\item{intfeature}{A binary vector indicating whether features are irrelevant (0) or relevant (1).}

\item{w}{A numerical vector of length 2 giving the prior probability of a variable being
irrelevant or relevant. The first slot is irrelevant the second relevant.}
}
\value{
The log PML
}
\description{
Computes the log PML for SUGS VarSel in the case of Gaussian mixtures
}
