% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compsugsStats.R
\name{compsugsStats}
\alias{compsugsStats}
\title{A function to compute statistics and posterior given clustering and priors}
\usage{
compsugsStats(n, nu_0, lambda_0, mu_0, S_0, member, X, numclust)
}
\arguments{
\item{n}{The vector indicating the number of observations in each cluster}

\item{nu_0}{The prior degrees of freedom}

\item{lambda_0}{The variance of the Guassian mean prior, the dafault value is \code{0.01}.}

\item{mu_0}{The prior mean}

\item{S_0}{The prior scale vector}

\item{member}{A vector of cluster membership allocaations}

\item{numclust}{The total number of occupied clusters}
}
\value{
The posterior hyperparameters for the normal inverse chisquared distribution
}
\description{
A function to compute statistics and posterior given clustering and priors
}
