% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compsugsStats.R
\name{compsugsStats1D}
\alias{compsugsStats1D}
\title{A function to compute statistics and posterior given clustering and priors}
\usage{
compsugsStats1D(n, nu_0, lambda_0, mu_0, S_0, member, mydata, numclust)
}
\arguments{
\item{n}{The vector indicating the number of observations in each cluster}

\item{nu_0}{The prior degrees of freedom}

\item{lambda_0}{The prior mean variance}

\item{mu_0}{The prior mean}

\item{S_0}{The prior scale vector}

\item{member}{A vector of cluster membership allocaations}

\item{mydata}{Data matrix with observations as rows.}

\item{numclust}{The total number of occupied clusters}
}
\value{
The posterior hyperparameters for the normal inverse chisquared distribution
}
\description{
A function to compute statistics and posterior given clustering and priors
}
