% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compvarMarg.R
\name{compvarMarg}
\alias{compvarMarg}
\title{A function to compute the marginal probability that of a variable of being relevant.}
\usage{
compvarMarg(K, D, n, lambda_0, nu_0, S_0, m, nu, lambda, S)
}
\arguments{
\item{K}{The number of currently occupied clusters}

\item{D}{The number of variables in the data matrix}

\item{n}{The vector indicating the number of observations in each cluster}

\item{lambda_0}{The prior mean variance}

\item{nu_0}{The prior degrees of freedom}

\item{S_0}{The prior scale vector}

\item{m}{The current posterior mean}

\item{nu}{The current posterior degrees of freedom}

\item{lambda}{The current posterior mean variance}

\item{S}{The current posterior scale vector}
}
\value{
The log marginal probability of a variable being relevant.
}
\description{
A function to compute the marginal probability that of a variable of being relevant.
}
