% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{This is a helper function for formatting data outputs from sugsvarsel}
\usage{
extract(X, results)
}
\arguments{
\item{X}{A data matrix with rows as observations.}

\item{results}{The output provided from runsugsvarsel}
}
\value{
A list including a vector of log marginal likelihoods, a matrix of memberships allocations, the random orderings
use in the algorithm a matrix of feature allocations.
}
\description{
This is a helper function for formatting data outputs from sugsvarsel
}
