% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isomix_documentation.R
\docType{data}
\name{Isomix}
\alias{Isomix}
\title{Small data example dataset with clustering structure}
\format{A list of length three. The data has 250 rows and 2 variables
\describe{
  \item{probIso}{The weights used to simulate the dataset}
  \item{samplesIso}{Numerical values with 250 rows and 2 variables}
  \item{componentsIso}{A vector of numerical value from 1 - 5 indicating cluster allocation}
}}
\usage{
data(Isomix)

A dataset containing 250 observations and 2 variables with 5 clusters.
}
\description{
Small data example dataset with clustering structure
}
\keyword{datasets}
