% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullView.R
\name{nullView}
\alias{nullView}
\title{A function to compute the marginal probability of belonging to the irrelevant partition.}
\usage{
nullView(X, D, N, lambda_0, nu_0, mu_0, S_0)
}
\arguments{
\item{X}{The data matrix with observations as rows.}

\item{D}{The number of variables of the data matrix.}

\item{N}{The number of observations.}

\item{lambda_0}{The prior value of the mean variance.}

\item{nu_0}{The prior number of degrees of freedom.}

\item{mu_0}{The prior mean.}

\item{S_0}{The prior scale for the inverse-chisquared distribution.}
}
\value{
Returns, in log form, the marginal probability of belong to the irrelevant
partition.
}
\description{
A function to compute the marginal probability of belonging to the irrelevant partition.
}
