% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma.R
\name{occamsWindow}
\alias{occamsWindow}
\title{Helper function to compute occams Window for SUGS VarSel}
\usage{
occamsWindow(ML, window)
}
\arguments{
\item{ML}{The vector of log MLs returned by sugsvarsel}

\item{window}{A numerical value indicating the size of occams window}
}
\value{
A numerical vector containg the normalised MLs.
}
\description{
Helper function to compute occams Window for SUGS VarSel
}
