% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phiUpdate.R
\name{phiUpdate}
\alias{phiUpdate}
\title{Function to update posterior for dirichlet concentration parameters.}
\usage{
phiUpdate(phi, i, clustnew, n, betaHat)
}
\arguments{
\item{phi}{Current value of posterior dirichlet weights.}

\item{i}{Current iteration of the SUGS algorithm.}

\item{clustnew}{The cluster to which the \code{i}th observation has allocated.}

\item{n}{A vector indicating the number of observations allocated to each cluster.}

\item{betaHat}{The grid prior for the dirichlet concentration parameter.}
}
\value{
An update value for phi.
}
\description{
Function to update posterior for dirichlet concentration parameters.
}
