% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runsugs1D.R
\name{runsugs1D}
\alias{runsugs1D}
\title{The SUGS clustering algorithm by Wang and Dunson (2011) for 1 dimensional data}
\usage{
runsugs1D(iter, mydata, Model)
}
\arguments{
\item{iter}{The number of random orderings of the observations for which to run
sugs}

\item{mydata}{Data matrix with observations as rows}

\item{Model}{Character string indicating whether to use PML, ML or
both for model selection. "Both" defaults to PML.}
}
\value{
A matrix of cluster allocation, K the number of clusters, a matrix
indicating the number of observations allocated to each cluster. The value of
the model selection criteria either log PML, log ML or both and the random orderings used
}
\description{
The SUGS clustering algorithm by Wang and Dunson (2011) for 1 dimensional data
}
