% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runsugsvarsel.R
\name{runsugsvarsel}
\alias{runsugsvarsel}
\title{Internal function for SUGS with variable selection algorithm. This function
manages the random orderings and loops.}
\usage{
runsugsvarsel(mydata, featiter, clustiter, intfeatures, Model, mu_0 = NULL,
  lambda_0 = 0.01, nu_0 = NULL, S_0 = NULL, betaHat = c(0.01, 0.1, 1, 5,
  10, 15, 30, 50, 100), a = 10, b = 1, w = c(0.5, 0.5),
  BPPARAM = bpparam())
}
\arguments{
\item{mydata}{A data matrix with rows as observations.}

\item{featiter}{The number of iterations of variable selection}

\item{clustiter}{The number of random ordering, for which to apply SUGS.}

\item{intfeatures}{A binary matrix of the intial variable set, probably chosen using function \code{pSelect}.
See documentation for \code{pSelect} for more details.}

\item{Model}{The method used for Model select, either PML, ML or Both. If you select both
the PML will be used to perform model selection.}

\item{mu_0}{The mean hyperparameter, default is the column means of the data matrix.}

\item{lambda_0}{The variance of the Guassian mean prior, the dafault value is \code{0.01}.}

\item{nu_0}{The degrees of freedom hyperparameter, the default value is \code{2 * (D + 2)}, where \code{D} is the number of variables.}

\item{S_0}{The scale hyperparamter, the deault value is a tenth of the column variance of the data matrix.}

\item{betaHat}{A grid of hyperparameters for the dirichlet concentration parameter, the default is \code{c(1, 5, 15, 30, 50, 100)}.}

\item{a}{The scale of the gamma prior for the dirichlet concentration parameter, the dafault value is \code{10}.}

\item{b}{The rate of the gamma prior for the dirichlet concentration parameter, the default value is \code{1}.}

\item{w}{The prior probability of a variable belong to the irrelevant or relevant partition. The vector must
contain two entries the first entry being the probabiliy of being irreleavnt and the second being the probability of being relevant
The default value is \code{c(0.5,0.5)}.}

\item{BPPARAM}{Support for parallel processing using the
\code{BiocParallel} infrastructure. When missing (default),
    the default registered \code{BiocParallelParam} parameters are
    used. Alternatively, one can pass a valid
    \code{BiocParallelParam} parameter instance: \code{SnowParam},
    \code{MulticoreParam}, \code{DoparParam}, \ldots see the
    \code{BiocParallel} package for details. To revert to the
    origianl serial implementation, use \code{NULL}.}
}
\value{
A matrix of cluster allocations for the number of iterations, a vector of either log PML, log ML or both,
the number of clusters at each interations, the random ordering used and the last function output for these.
}
\description{
Internal function for SUGS with variable selection algorithm. This function
manages the random orderings and loops.
}
