% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleMix.R
\name{sampleMix}
\alias{sampleMix}
\title{Generate data from a mixture of normals}
\usage{
sampleMix(N, weights, mean, cov)
}
\arguments{
\item{N}{The number of points to be generated}

\item{weights}{A vector of weights for each mixture component}

\item{mean}{A list containing the means of each component}

\item{cov}{A list of covariance matricies for each component}
}
\value{
A list returning the weights, the sampled values and the component allocations.
}
\description{
Generate data from a mixture of normals
}
