% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleMix.R
\name{samplemixDiag}
\alias{samplemixDiag}
\title{Generate data from a mixture of normals with diagonal structure}
\usage{
samplemixDiag(N, weights, mean, cov)
}
\arguments{
\item{N}{The number of points to be generated}

\item{weights}{A vector of weights for each mixture component}

\item{mean}{A list containing the means of each component}

\item{cov}{A list of covariance as a vector for each component}
}
\value{
A list returning the weights, the sampled values and the component allocations.
}
\description{
Generate data from a mixture of normals with diagonal structure
}
