% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleMix.R
\name{samplemixNoise}
\alias{samplemixNoise}
\title{Generate data from a mixture of normals with additionally noisy variables}
\usage{
samplemixNoise(N, weights, mean, cov, p, meanNoise, covNoise, struc = "Diag")
}
\arguments{
\item{N}{The number of observations to be generated}

\item{weights}{A vector of weights for each mixture component}

\item{mean}{A list containing the means for each component}

\item{cov}{A list containing covariance matrix or variance vectors. See paramater \code{struc}}

\item{p}{The number of noisy variables}

\item{meanNoise}{A vector of the means of each of the noisy variables}

\item{covNoise}{A vector of standard deviations}

\item{struc}{A character string indicating the structure of the covariance matrix. Diagonal is selected by default
and \code{cov} should be provided as a list of vectors. If "Full" is specified, then \code{cov} should be provided
as list of covariance matricies.}
}
\value{
A list of three of length three. The first the mixture weights,
the second a data matrix containing the observations, and the third a vector of the componenet
allocations.
}
\description{
Generate data from a mixture of normals with additionally noisy variables
}
