% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugs1D.R
\name{sugs1D}
\alias{sugs1D}
\title{The SUGS clustering algorithm by Wang and Dunson (2011) for 1 dimensional data}
\usage{
sugs1D(mydata, Model)
}
\arguments{
\item{mydata}{Data matrix with observations as rows}

\item{Model}{Character string indicating whether to use PML, ML or
both for model selection. "Both" defaults to PML.}
}
\value{
A matrix of cluster allocation, K the number of clusters, a matrix
indicating the number of observations allocated to each cluster. The value of
the model selection criteria either log PML, log ML or both.
}
\description{
The SUGS clustering algorithm by Wang and Dunson (2011) for 1 dimensional data
}
