% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugsclusterProb.R
\name{sugsclusterProb}
\alias{sugsclusterProb}
\title{A function to compute the marginal probability of belong to an occupied cluster for the SUGS algorithm}
\usage{
sugsclusterProb(x, K, i, D, n, betaHat, phi, m, nu, S, lambda)
}
\arguments{
\item{x}{The current observation under consideration.}

\item{K}{The number of occupied clusters}

\item{i}{The current iteration of the SUGS algorithm}

\item{D}{The number of variables}

\item{n}{A numeric vector containing the number of observations allocated to each cluster}

\item{betaHat}{The grid of values for the dirichlet concentration parameter}

\item{phi}{The weights associated with the dirichlet grid prior}

\item{m}{The current posterior mean}

\item{nu}{The current posterior degrees of freedom}

\item{S}{The current posterior scale vector}

\item{lambda}{The current posterior for the mean variance}
}
\value{
The numeric vector with the unnormalised probability of beloging to each cluster.
}
\description{
A function to compute the marginal probability of belong to an occupied cluster for the SUGS algorithm
}
