% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugsclustMarg.R
\name{sugsclustMarg}
\alias{sugsclustMarg}
\title{Computes the marginal probability of an observation belonging to a currently occupied cluster.}
\usage{
sugsclustMarg(x, K, i, D, n, betaHat, phi, m, nu, S, lambda, intfeature)
}
\arguments{
\item{x}{The observation currently underconsideration.}

\item{K}{The number of clusters.}

\item{i}{The current iteration state of the algorithm.}

\item{D}{The number of variable in the data matrix.}

\item{n}{The vector indicating the number of observations in each cluster}

\item{betaHat}{The grid of concentration parameters.}

\item{phi}{The probability weights of \code{beteHat}.}

\item{m}{The current posterior mean.}

\item{nu}{The current posterior degrees of freedom.}

\item{S}{The current posterior scale vector.}

\item{lambda}{The current posterior mean variance.}

\item{intfeature}{A binary vector indicating whether features are irrelevant (0) or relevant (1).}
}
\value{
The unnormalised probability of belong to any of the currently occupied clusters.
}
\description{
Computes the marginal probability of an observation belonging to a currently occupied cluster.
}
