% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugscompMl.R
\name{sugscompMl}
\alias{sugscompMl}
\title{Computes the log marginal likelihood of clustering for vanilla SUGS}
\usage{
sugscompMl(K, D, n, nu_0, S_0, lambda_0, nu, lambda, S)
}
\arguments{
\item{K}{The number of currently occupied clusters}

\item{D}{The number of variable in the data}

\item{n}{The vector indicating the number of observations in each cluster}

\item{nu_0}{The prior degrees of freedom}

\item{S_0}{The prior scale vector}

\item{lambda_0}{The prior mean variance}

\item{nu}{The current posterior degrees of freedom}

\item{lambda}{The current posterior mean variance}

\item{S}{The current posterior scale vector}
}
\value{
The log Marginal likelihood of the clustering
}
\description{
Computes the log marginal likelihood of clustering for vanilla SUGS
}
