% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugscompPml.R
\name{sugscompPml}
\alias{sugscompPml}
\title{Function to compute the PML for vanilla SUGS}
\usage{
sugscompPml(X, K, N, D, n, phi, betaHat, m, nu, lambda, S, mu_0, nu_0, lambda_0,
  S_0)
}
\arguments{
\item{X}{The data matrix with rows as observations}

\item{K}{The number of currently occupied clusters}

\item{D}{The number of variables}

\item{n}{The vector indicating the number of observations in each cluster}

\item{phi}{The weights associated with the dirichlet grid prior}

\item{betaHat}{The grid of values for the dirichlet concentration parameter}

\item{m}{The current posterior mean}

\item{nu}{The current posterior degrees of freedom}

\item{lambda}{The current posterior mean variance}

\item{S}{The current posterior scale vector}

\item{mu_0}{The prior mean}

\item{nu_0}{The prior degrees of freedom}

\item{S_0}{The prior scale vector}
}
\value{
The log PML.
}
\description{
Function to compute the PML for vanilla SUGS
}
