% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sugscompPmlcpp}
\alias{sugscompPmlcpp}
\title{A C++ accelerated version of sugscompPml}
\usage{
sugscompPmlcpp(X, K, N, D, n, phi, betaHat, m, nu, lambda, S, mu_0, nu_0,
  lambda_0, S_0)
}
\arguments{
\item{X}{The data matrix with rows as observations}

\item{K}{An integer specifiying the number of clusters}

\item{D}{An integer specifiying the number of variables}

\item{n}{A numeric vector contain the number of people already allocated to each cluster}

\item{phi}{A numeric matrix containing the weights for the dirichlet hyperpiors}

\item{betaHat}{A numeric vector containing the grid of hyperpriors for the dirichlet
concentration parameter}

\item{m}{A numeric matrix containing the means for each component}

\item{nu}{A numeric vector containg the degrees of freedom for each component}

\item{lambda}{A numeric vector containing the mean variance hyperparamter for each component}

\item{S}{A numeric matrix containing the scale prior for each component}

\item{mu_0}{The mean hyperparameter, default is the column means of the data matrix.}

\item{nu_0}{The degrees of freedom hyperparameter, the default value is \code{2 * (D + 2)}, where \code{D} is the number of variables.}

\item{lambda_0}{The variance of the Guassian mean prior, the dafault value is \code{0.01}.}

\item{S_0}{The scale hyperparamter, the default value is a tenth of the column variance of the data matrix.}
}
\value{
The log PML.
}
\description{
A C++ accelerated version of sugscompPml
}
