% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugsclusterProb.R
\name{sugsnewclusterProb}
\alias{sugsnewclusterProb}
\title{A function to compute the probability of belonging to an unoccupied cluster for the SUGS algorithm}
\usage{
sugsnewclusterProb(x, i, D, phi, betaHat, mu_0, lambda_0, nu_0, S_0)
}
\arguments{
\item{x}{The current observation under consideration.}

\item{i}{The current iteration of the SUGS algorithm}

\item{D}{The number of variables}

\item{phi}{The weights associated with the dirichlet grid prior}

\item{betaHat}{The grid of values for the dirichlet concentration parameter}
}
\value{
The unnormlised probability of the observation belonging to a new cluster
}
\description{
A function to compute the probability of belonging to an unoccupied cluster for the SUGS algorithm
}
