% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugsclustMarg.R
\name{sugsnewclustMarg}
\alias{sugsnewclustMarg}
\title{Computes the marginal probability of an observation belonging to a currently unoccupied cluster.}
\usage{
sugsnewclustMarg(x, i, D, phi, betaHat, mu_0, lambda_0, nu_0, S_0, intfeatures)
}
\arguments{
\item{x}{The observation currently underconsideration.}

\item{i}{The current iteration state of the algorithm.}

\item{D}{The number of variable in the data matrix.}

\item{phi}{The probability weights of \code{beteHat}.}

\item{betaHat}{The grid of concentration parameters.}

\item{mu_0}{The prior mean.}

\item{lambda_0}{The prior mean variance.}

\item{nu_0}{The prior degrees of freedom.}

\item{S_0}{The prior scale vector.}

\item{intfeatures}{A binary vector indicating whether features are irrelevant (0) or relevant (1).}
}
\value{
The unnormalised probability of belong to any of the currently unoccupied clusters.
}
\description{
Computes the marginal probability of an observation belonging to a currently unoccupied cluster.
}
