% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sugsnewclustMargcpp}
\alias{sugsnewclustMargcpp}
\title{A C++ accelerated version of sugsnewclustMarg}
\usage{
sugsnewclustMargcpp(x, i, D, phi, betaHat, mu_0, nu_0, S_0, lambda_0,
  intfeature)
}
\arguments{
\item{x}{A numeric vector containing data of the observation}

\item{i}{An integer specifying the current iteration of SUGS algorithm}

\item{D}{An integer specifiying the number of variables}

\item{phi}{A numeric matrix containing the weights for the dirichlet hyperpiors}

\item{betaHat}{A numeric vector containing the grid of hyperpriors for the dirichlet
concentration parameter}

\item{mu_0}{The mean hyperparameter, default is the column means of the data matrix.}

\item{nu_0}{The degrees of freedom hyperparameter, the default value is \code{2 * (D + 2)}, where \code{D} is the number of variables.}

\item{S_0}{The scale hyperparamter, the default value is a tenth of the column variance of the data matrix.}

\item{lambda_0}{The variance of the Guassian mean prior, the dafault value is \code{0.01}.}

\item{intfeature}{A binary vector of feature which are parition as irrelevant (0) or relevant (1).}
}
\value{
The unnormalised probability of belong to any of the currently unoccupied clusters.
}
\description{
A C++ accelerated version of sugsnewclustMarg
}
