% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugsvarAlloc.R
\name{sugsvarAlloc}
\alias{sugsvarAlloc}
\title{A function to partition the variable into relevant and irrelevant}
\usage{
sugsvarAlloc(D, varMarg, lognullMarg, intfeature, w)
}
\arguments{
\item{D}{The number of variables in the data matrix}

\item{varMarg}{The log marginal probability of a cluster being relevant}

\item{lognullMarg}{The log marginal probability of a variable belonging to the null model}

\item{intfeature}{A binary vector indicating whether features are irrelevant (0) or relevant (1).}

\item{w}{A numerical vector of length 2 giving the prior probability of a variable being
irrelevant or relevant. The first slot is irrelevant the second relevant.}
}
\value{
A binary vector indicating whether a variable is relevant (1) or irrelevant (0)
}
\description{
A function to partition the variable into relevant and irrelevant
}
