% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugsclusterProb.R
\name{sugsnewclusterProb}
\alias{sugsnewclusterProb}
\title{A function to compute the probability of belonging to an unoccupied cluster for the SUGS algorithm}
\usage{
sugsnewclusterProb(x, i, D, phi, betaHat, mu_0, lambda_0, nu_0, S_0)
}
\arguments{
\item{x}{The current observation under consideration.}

\item{i}{The current iteration of the SUGS algorithm}

\item{D}{The number of variables}

\item{phi}{The weights associated with the dirichlet grid prior}

\item{betaHat}{The grid of values for the dirichlet concentration parameter}

\item{mu_0}{The mean hyperparameter, default is the column means of the data matrix.}

\item{lambda_0}{The variance of the Guassian mean prior, the dafault value is \code{0.01}.}

\item{nu_0}{The degrees of freedom hyperparameter, the default value is \code{D}, where \code{D} is the number of variables.}

\item{S_0}{The scale hyperparamter, the deault value is a fifth of the column variance of the data matrix.}
}
\value{
The unnormlised probability of the observation belonging to a new cluster
}
\description{
A function to compute the probability of belonging to an unoccupied cluster for the SUGS algorithm
}
